Fix Connectivity
When losing connection and reconnecting, even though the device is connected
it sometimes needs some time before it can ping 8.8.8.8 successfully.
Because of this a lot of time the resolution is "no connection" even
though device is connected.

Fix:
When receiving connectivity change, check connectivity state and only if connected
try pinging 8.8.8.8. If ping fails, retry every 1.5sec upto 5 times